+++ app/src/org/commcare/tasks/ResourceRecoveryTask.java
@@ -9,6 +9,7 @@ import org.commcare.dalvik.R;
 import org.commcare.engine.resource.AppInstallStatus;
 import org.commcare.engine.resource.ResourceInstallUtils;
 import org.commcare.network.CommcareRequestGenerator;
+import org.commcare.network.RequestStats;
 import org.commcare.resources.model.InstallCancelled;
 import org.commcare.resources.model.InstallCancelledException;
 import org.commcare.resources.model.InstallRequestSource;
