+++ app/src/org/commcare/update/UpdateTask.java
@@ -5,7 +5,7 @@ import android.content.Context;
 import org.commcare.engine.resource.AppInstallStatus;
 import org.commcare.resources.model.InstallCancelled;
 import org.commcare.resources.model.InstallRequestSource;
-import org.commcare.tasks.RequestStats;
+import org.commcare.network.RequestStats;
 import org.commcare.tasks.ResultAndError;
 import org.commcare.tasks.SingletonTask;
