+++ app/src/org/commcare/update/UpdateWorker.kt
@@ -12,7 +12,7 @@ import org.commcare.engine.resource.AppInstallStatus
 import org.commcare.engine.resource.ResourceInstallUtils
 import org.commcare.resources.model.InstallCancelled
 import org.commcare.resources.model.InstallRequestSource
-import org.commcare.tasks.RequestStats
+import org.commcare.network.RequestStats
 import org.commcare.tasks.ResultAndError
 
 /**
